CargasAPI hecho (en teoria), falta dejar ejemplos de curl y probar que anden todos bien

+++ src/main/java/org/tallerjava/moduloCargas/interfase/remota/rest/CargaIniciarDTO.java
@@ -11,5 +11,5 @@ import org.tallerjava.moduloCargas.dominio.*;
 public class CargaIniciarDTO {
     private int idCargador;
     private String cedula;
-    private int idMedioPago;
+    private long idMedioPago;
 }
